refactor(utils): move openUrl to helpers
