+++ core/src/utils/theme.ts
@@ -1,5 +1,4 @@
-import type { RouterDirection } from '../components/router/utils/interface';
-import type { AnimationBuilder, Color, CssClassMap } from '../interface';
+import type { Color, CssClassMap } from '../interface';
 
 export const hostContext = (selector: string, el: HTMLElement): boolean => {
   return el.closest(selector) !== null;
@@ -33,24 +32,4 @@ export const getClassMap = (classes: string | string[] | undefined): CssClassMap
   const map: CssClassMap = {};
   getClassList(classes).forEach((c) => (map[c] = true));
   return map;
-};
-
-const SCHEME = /^[a-z][a-z0-9+\-.]*:/;
-
-export const openURL = async (
-  url: string | undefined | null,
-  ev: Event | undefined | null,
-  direction: RouterDirection,
-  animation?: AnimationBuilder
-): Promise<boolean> => {
-  if (url != null && url[0] !== '#' && !SCHEME.test(url)) {
-    const router = document.querySelector('ion-router');
-    if (router) {
-      if (ev != null) {
-        ev.preventDefault();
-      }
-      return router.push(url, direction, animation);
-    }
-  }
-  return false;
 };
